refactor: split input dispatch out of main.rs
`main.rs` was 537 lines, mostly `handle_key_*` / `handle_mouse` /
`apply_horizontal_scroll` — every tab's keyboard semantics crammed next
to terminal setup code.

Extracted to `src/input.rs` (435 lines):
- `pub fn handle_key(key, app)` — global shortcuts + tab dispatch
- private `handle_key_git / handle_key_files / handle_key_graph`
- `pub fn handle_mouse<B: Backend>(mouse, app, terminal)`
- private `apply_horizontal_scroll` helper
- `pub const DOUBLE_CLICK_WINDOW`

`main.rs` is now 118 lines: terminal bootstrap (panic hook, raw mode,
mouse capture), the main loop, and the event-drain loop. Two inputs
stay inlined in main.rs:

- `v` (select mode toggle) — needs `execute!(terminal.backend_mut(),
  Disable/EnableMouseCapture)` which can't easily go through `&Terminal<B>`
  alone.
- `show_help` dismiss — one state flip, doesn't pay for its own module.

No behavior change. 124/124 tests, clippy + fmt clean.
